// types/githubEvents.ts
export interface GithubEvent {
  id: string
  type: string
  actor: Actor
  repo: Repo
  payload: Payload
  public: boolean
  created_at: string
  org?: Org
}

export interface Actor {
  id: number
  login: string
  display_login: string
  gravatar_id: string
  url: string
  avatar_url: string
}

export interface Repo {
  id: number
  name: string
  url: string
}

export interface Payload {
  action?: string
  ref?: string
  ref_type?: string
  pusher_type?: string
  push_id?: number
  size?: number
  distinct_size?: number
  head?: string
  before?: string
  commits?: Commit[]
  number?: number
  pull_request?: PullRequest
  master_branch?: string
  description?: string
}

export interface Commit {
  sha: string
  author: Author
  message: string
  distinct: boolean
  url: string
}

export interface Author {
  email: string
  name: string
}

export interface PullRequest {
  url: string
  id: number
  node_id: string
  html_url: string
  diff_url: string
  patch_url: string
  issue_url: string
  number: number
  state: string
  locked: boolean
  title: string
  user: User
  body?: string
  created_at: string
  updated_at: string
  closed_at?: string
  merged_at?: string
  merge_commit_sha?: string
  assignee?: Assignee
  assignees: Assignee2[]
  requested_reviewers: any[]
  requested_teams: any[]
  labels: Label[]
  milestone: any
  draft: boolean
  commits_url: string
  review_comments_url: string
  review_comment_url: string
  comments_url: string
  statuses_url: string
  head: Head
  base: Base
  _links: Links
  author_association: string
  auto_merge: any
  active_lock_reason: any
  merged: boolean
  mergeable: any
  rebaseable: any
  mergeable_state: string
  merged_by?: MergedBy
  comments: number
  review_comments: number
  maintainer_can_modify: boolean
  commits: number
  additions: number
  deletions: number
  changed_files: number
}
// ...
export interface Org {
  id: number
  login: string
  gravatar_id: string
  url: string
  avatar_url: string
}